+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This version brings version updates for cluster infrastructure components (Nomad
 ### Upgrade Notes
 - Follow ["clean reset" procedure](https://github.com/liquidinvestigations/docs/wiki/Maintenance#clean-reset) with **[cluster version 0.16.0](https://github.com/liquidinvestigations/cluster/tree/v0.16.0)**
 
+### Improvements
+- Deployment: The `liquid deploy` command now better shows all service error outputs in one place.
 ---